fix so it runs on clusters (where there is no weekuserusage table)
git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@2383 1dfba125-df10-0410-8a77-af1a8d6d074f

+++ bin/upgrading/update-db-general.pl
@@ -1798,7 +1798,8 @@ register_alter(sub {
                  "ALTER TABLE community DROP ownerid");
     }
 
-    unless (column_type("weekuserusage", "uafter")) {
+    # if it exists, but it's the old way, just kill it.
+    if (column_type("weekuserusage", "ubefore") && ! column_type("weekuserusage", "uafter")) {
         do_sql("DROP TABLE weekuserusage");
         create_table("weekuserusage");
     }
